Rename from SkrivTilBehandler to MeldingTilBehandler
Overskriften i selve panelet er "Skriv til behandler", men komponenten trenger ikke å hete dette. MeldingTilBehandler er mer aligned med navnbruken vi har i backend.
